pyproject.toml
[tool.poetry]
name = "radicalbit-ai-monitoring"
# x-release-please-start-version
version = "0.1.0"
# x-release-please-end
description = "Radicalbit AI Monitoring platform installer"
authors = ["Radicalbit"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
pydantic = "^2.7.4"
pydantic-settings = "^2.3.4"
requests = "^2.32.3"
pyfiglet = "^1.0.2"
pyyaml = "^6.0.1"
yaspin = "^3.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pyyaml = "^6.0.1"
ruff = "^0.4.10"

[tool.poetry.scripts]
rbit-ai-monitoring = "radicalbit_ai_monitoring.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
